requirements.txt
pandas>=2.2.2
numpy>=1.26.4
pyarrow>=21.0.0
openpyxl>=3.1.2
PyYAML>=6.0.3
